+++ tests/unit/testdisparity_graph.hpp
@@ -1,0 +1,10 @@
+#include "gtest/gtest.h"
+#include "disparity_graph.hpp"
+
+class DisparityGraphTest : public ::testing::Test {
+    protected:
+        DisparityGraphTest() {};
+        virtual ~DisparityGraphTest() {};
+        virtual void SetUp() {};
+        virtual void TearDown() {};
+};